tweak(base): fix com pot value

+++ arma/base.lua
@@ -330,7 +330,7 @@ Citizen.CreateThread(function()
     MySQL.SingleQuery([[
     CREATE TABLE IF NOT EXISTS arma_community_pot (
     arma VARCHAR(65) NOT NULL,
-    value INT(11) NOT NULL,
+    value BIGINT(11) NOT NULL,
     PRIMARY KEY (arma)
     );
     ]])
